clyde: automated changes on clyde/7c1e80b7/FinanceAppDesktop

+++ repository_pattern/expenses_repo.py
@@ -4,6 +4,8 @@ from database.engine import SessionLocal
 
 
 class ExpenseRepository(RepositoryInterface):
+    ALLOWED_UPDATE_FIELDS = frozenset({"expenses_category", "expenses_money", "expenses_date"})
+
     def __init__(self):
         self.session = SessionLocal()
 
@@ -15,12 +17,19 @@ class ExpenseRepository(RepositoryInterface):
         return self.session.query(Expense).all()
 
     def update(self, entity_id, **kwargs):
+        for key in kwargs:
+            if key not in self.ALLOWED_UPDATE_FIELDS:
+                raise ValueError(f"Field '{key}' is not allowed to be updated")
         obj = self.session.get(Expense, entity_id)
+        if obj is None:
+            raise LookupError(f"Expense with id {entity_id} not found")
         for key, value in kwargs.items():
             setattr(obj, key, value)
         self.session.commit()
 
     def delete(self, entity_id):
-        obj = self.session.query(Expense).get(entity_id)
+        obj = self.session.get(Expense, entity_id)
+        if obj is None:
+            raise LookupError(f"Expense with id {entity_id} not found")
         self.session.delete(obj)
         self.session.commit()
